Saving work in progress
Saving work in progress

saving work

Have inspector view registry to show a view in the inspection sidebar when a inspector view provider exists.

Add inspector view registry

Get the composition for the selected object to show in the Elements pool

Show an appropriate message if compostion is empty.

Select the scope domain object when clearing a selection.